Markdown components && half screen form
