+++ app/src/pages/Create.js
@@ -3,14 +3,14 @@ import Box from "@material-ui/core/Box";
 import Typography from "@material-ui/core/Typography";
 import Alert from '@material-ui/lab/Alert';
 import Debug from "debug";
-import Markdown from 'markdown-to-jsx';
 import React, { useCallback, useMemo } from "react";
 import FormView from '../components/Form';
 import { SEO } from "../components/Helmet";
 import NotebookTitle from "../components/NotebookTitle";
 import { getNotebookMetadata } from "../utils/notebookMetadata";
 
-
+import NotebookImage from '../components/organisms/markdownParsers/NotebookImage';
+import NotebookInfo from '../components/organisms/markdownParsers/NotebookInfo';
 
 const debug = Debug("Create");
 
@@ -35,25 +35,38 @@ export default React.memo(function Create({ ipfs, node, dispatch }) {
       <NotebookTitle name={metadata?.name} />
       <AlertMessage connected={connected}/>
 
+        <TwoColumns>
+          {/* FORM INPUTS */}
+          <div>
+            <Typography variant="h5" gutterBottom>
+              Inputs
+            </Typography>
 
-        {/* FORM INPUTS */}
+            <FormView
+              input={ipfs?.input}
+              connected={connected}
+              metadata={metadata}
+              onSubmit={dispatch}
+            />
+          </div> 
+
+          {/* OUTPUTS */}
+          <div>
+            <Typography variant="h5" gutterBottom>
+              Output
+            </Typography>
+            <NotebookImage metadata={metadata} style={{width: '50%', padding: '1.5em 0'}}/> 
+          </div>  
+        </TwoColumns>
+
+        {/* NOTEBOOK DESCRIPTION */}
         <div>
           <Typography variant="h5" gutterBottom>
-            Inputs
+          Details
           </Typography>
-
-          <FormView
-            input={ipfs?.input}
-            connected={connected}
-            metadata={metadata}
-            onSubmit={dispatch}
-          />
-        </div>      
-
-        <br/><br/><br/>
-        {/* NOTEBOOK DESCRIPTION */}
-        <NotebookDescription metadata={metadata} />
-
+          <NotebookInfo noImg description={metadata.description}/>
+        </div>
+          
     </Box>
 });
 
@@ -66,24 +79,10 @@ width: 100%;
 margin-top: 1em;
 `
 
-// Notebook Description
-const NotebookDescription = ({ metadata }) => {
-  if (metadata === null) return <></>
-  return <Box >
-    <Typography variant="h5" gutterBottom>
-        Details
-    </Typography>
-    <Typography color="textSecondary">
-      <Markdown children={metadata.description} />
-    </Typography>
-  </Box>
-}
-
-
 // Alert Message
 const AlertMessage = ({ connected }) => {
   if (connected) return <></>
-  return <Alert severity="info">
+  return <Alert severity="info" style={{margin: '2em 0'}}>
     The inputs are <b>disabled</b> because <b>no Colab node is running</b>
     ! Click on <b>LAUNCH</b> (bottom right) or refer to INSTRUCTIONS for further instructions.
   </Alert>
